services/orion-notify/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
SQLAlchemy==2.0.43
pyyaml==6.0.2
